rename profile_features remove min/max
